Refatorando asgard.math.round_up para retornar Decimal

+++ asgard/math/__init__.py
@@ -1,5 +1,5 @@
 from decimal import Decimal, ROUND_UP
 
 
-def round_up(n: Decimal, prec: int = 2) -> float:
-    return float(n.quantize(Decimal("." + "0" * prec), rounding=ROUND_UP))
+def round_up(n: Decimal, prec: int = 2) -> Decimal:
+    return n.quantize(Decimal("." + "0" * prec), rounding=ROUND_UP)
